+++ CHANGELOG.md
@@ -2,6 +2,9 @@
 
 ## Unreleased
 
+- Add `--bind` and `--lan` options to `voom forward auto enable` so automatic
+  forwards can be exposed on a chosen host interface instead of loopback only.
+
 ## v1.0.0 - 2026-06-01
 
 - Initial release.